Add find_file_offset() method to ElfParser
This change moves the file offset retrieval logic from the ELF resolver
into the ELF parser, where it will be better co-located with similar
functionality and is generally more accessible.
